Added AjoutercriditServlet updated Ajoutercredit.jsp file && Dashboard.jsp && Credit.java

+++ src/main/java/Models/Credit.java
@@ -6,9 +6,9 @@ public class Credit {
     private int id_credit ;
     private String nom;
     private String description;
-    private Double montant;
+    private int montant;
     private int duree;
-    private Double taux;
+    private int taux;
     private String simulation;
     private Date date_creation;
 
@@ -36,11 +36,11 @@ public class Credit {
         this.description = description;
     }
 
-    public Double getMontant() {
+    public int getMontant() {
         return montant;
     }
 
-    public void setMontant(Double montant) {
+    public void setMontant(int montant) {
         this.montant = montant;
     }
 
@@ -52,11 +52,11 @@ public class Credit {
         this.duree = duree;
     }
 
-    public Double getTaux() {
+    public int getTaux() {
         return taux;
     }
 
-    public void setTaux(Double taux) {
+    public void setTaux(int taux) {
         this.taux = taux;
     }
 
@@ -77,7 +77,7 @@ public class Credit {
     }
     public Credit(){}
 
-    public Credit(int id_credit, String nom, String description, Double montant, int duree, Double taux, String simulation, Date date_creation) {
+    public Credit(int id_credit, String nom, String description, int montant, int duree, int taux, String simulation, Date date_creation) {
         this.id_credit = id_credit;
         this.nom = nom;
         this.description = description;
